requirements.txt
Click==7.0
Flask==1.0.2
Flask-WTF==0.14.2
gunicorn==19.9.0
ItsDangerous==1.0.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.3.0
pytz==2018.5
Werkzeug==0.14.1
WTForms==2.2.1
